students_part_2/MyProject/fastapi/requirements.txt
fastapi
uvicorn
mysql-connector-python
matplotlib
geopandas
